# perf: Faster decimal precision overflow checks