--- a/arrow-array/benches/decimal_overflow.rs
+++ b/arrow-array/benches/decimal_overflow.rs
@@ -15,26 +15,37 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use arrow_array::builder::Decimal128Builder;
+use arrow_array::builder::{Decimal128Builder, Decimal256Builder};
+use arrow_buffer::i256;
 use criterion::*;
 
 fn criterion_benchmark(c: &mut Criterion) {
     let len = 8192;
-    let mut b = Decimal128Builder::with_capacity(len);
+    let mut builder_128 = Decimal128Builder::with_capacity(len);
+    let mut builder_256 = Decimal256Builder::with_capacity(len);
     for i in 0..len {
         if i % 10 == 0 {
-            b.append_value(i128::max_value());
+            builder_128.append_value(i128::max_value());
+            builder_256.append_value(i256::from_i128(i128::max_value()));
         } else {
-            b.append_value(i as i128);
+            builder_128.append_value(i as i128);
+            builder_256.append_value(i256::from_i128(i as i128));
         }
     }
-    let array = b.finish();
+    let array_128 = builder_128.finish();
+    let array_256 = builder_256.finish();
 
-    c.bench_function("validate_decimal_precision", |b| {
-        b.iter(|| black_box(array.validate_decimal_precision(8).unwrap_or(())));
+    c.bench_function("validate_decimal_precision_128", |b| {
+        b.iter(|| black_box(array_128.validate_decimal_precision(8).unwrap_or(())));
     });
-    c.bench_function("null_if_overflow_precision", |b| {
-        b.iter(|| black_box(array.null_if_overflow_precision(8)));
+    c.bench_function("null_if_overflow_precision_128", |b| {
+        b.iter(|| black_box(array_128.null_if_overflow_precision(8)));
+    });
+    c.bench_function("validate_decimal_precision_256", |b| {
+        b.iter(|| black_box(array_256.validate_decimal_precision(8).unwrap_or(())));
+    });
+    c.bench_function("null_if_overflow_precision_256", |b| {
+        b.iter(|| black_box(array_256.null_if_overflow_precision(8)));
     });
 }