--- a/arrow-data/src/decimal.rs
+++ b/arrow-data/src/decimal.rs
@@ -23,10 +23,13 @@ pub use arrow_schema::{
     DECIMAL_DEFAULT_SCALE,
 };
 
-// MAX decimal256 value of little-endian format for each precision.
-// Each element is the max value of signed 256-bit integer for the specified precision which
-// is encoded to the 32-byte width format of little-endian.
-pub(crate) const MAX_DECIMAL_BYTES_FOR_LARGER_EACH_PRECISION: [i256; 76] = [
+/// MAX decimal256 value of little-endian format for each precision.
+/// Each element is the max value of signed 256-bit integer for the specified precision which
+/// is encoded to the 32-byte width format of little-endian.
+/// The first element is unused and is inserted so that we can look up using
+/// precision as the index without the need to subtract 1 first.
+pub(crate) const MAX_DECIMAL_BYTES_FOR_LARGER_EACH_PRECISION: [i256; 77] = [
+    i256::from_i128(0_i128), // unused first element
     i256::from_le_bytes([
         9, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
         0, 0,
@@ -333,10 +336,13 @@ pub(crate) const MAX_DECIMAL_BYTES_FOR_LARGER_EACH_PRECISION: [i256; 76] = [
     ]),
 ];
 
-// MIN decimal256 value of little-endian format for each precision.
-// Each element is the min value of signed 256-bit integer for the specified precision which
-// is encoded to the 76-byte width format of little-endian.
-pub(crate) const MIN_DECIMAL_BYTES_FOR_LARGER_EACH_PRECISION: [i256; 76] = [
+/// MIN decimal256 value of little-endian format for each precision.
+/// Each element is the min value of signed 256-bit integer for the specified precision which
+/// is encoded to the 76-byte width format of little-endian.
+/// The first element is unused and is inserted so that we can look up using
+/// precision as the index without the need to subtract 1 first.
+pub(crate) const MIN_DECIMAL_BYTES_FOR_LARGER_EACH_PRECISION: [i256; 77] = [
+    i256::from_i128(0_i128), // unused first element
     i256::from_le_bytes([
         247, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255,
         255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255,
@@ -644,8 +650,11 @@ pub(crate) const MIN_DECIMAL_BYTES_FOR_LARGER_EACH_PRECISION: [i256; 76] = [
 ];
 
 /// `MAX_DECIMAL_FOR_EACH_PRECISION[p]` holds the maximum `i128` value that can
-/// be stored in [arrow_schema::DataType::Decimal128] value of precision `p`
-pub const MAX_DECIMAL_FOR_EACH_PRECISION: [i128; 38] = [
+/// be stored in [arrow_schema::DataType::Decimal128] value of precision `p`.
+/// The first element is unused and is inserted so that we can look up using
+/// precision as the index without the need to subtract 1 first.
+pub(crate) const MAX_DECIMAL_FOR_EACH_PRECISION: [i128; 39] = [
+    0, // unused first element
     9,
     99,
     999,
@@ -687,8 +696,11 @@ pub const MAX_DECIMAL_FOR_EACH_PRECISION: [i128; 38] = [
 ];
 
 /// `MIN_DECIMAL_FOR_EACH_PRECISION[p]` holds the minimum `i128` value that can
-/// be stored in a [arrow_schema::DataType::Decimal128] value of precision `p`
-pub const MIN_DECIMAL_FOR_EACH_PRECISION: [i128; 38] = [
+/// be stored in a [arrow_schema::DataType::Decimal128] value of precision `p`.
+/// The first element is unused and is inserted so that we can look up using
+/// precision as the index without the need to subtract 1 first.
+pub(crate) const MIN_DECIMAL_FOR_EACH_PRECISION: [i128; 39] = [
+    0, // unused first element
     -9,
     -99,
     -999,
@@ -738,16 +750,15 @@ pub fn validate_decimal_precision(value: i128, precision: u8) -> Result<(), Arro
             "Max precision of a Decimal128 is {DECIMAL128_MAX_PRECISION}, but got {precision}",
         )));
     }
-    let idx = usize::from(precision) - 1;
-    if value > MAX_DECIMAL_FOR_EACH_PRECISION[idx] {
+    if value > MAX_DECIMAL_FOR_EACH_PRECISION[precision as usize] {
         Err(ArrowError::InvalidArgumentError(format!(
             "{value} is too large to store in a Decimal128 of precision {precision}. Max is {}",
-            MAX_DECIMAL_FOR_EACH_PRECISION[idx]
+            MAX_DECIMAL_FOR_EACH_PRECISION[precision as usize]
         )))
-    } else if value < MIN_DECIMAL_FOR_EACH_PRECISION[idx] {
+    } else if value < MIN_DECIMAL_FOR_EACH_PRECISION[precision as usize] {
         Err(ArrowError::InvalidArgumentError(format!(
             "{value} is too small to store in a Decimal128 of precision {precision}. Min is {}",
-            MIN_DECIMAL_FOR_EACH_PRECISION[idx]
+            MIN_DECIMAL_FOR_EACH_PRECISION[precision as usize]
         )))
     } else {
         Ok(())
@@ -758,10 +769,9 @@ pub fn validate_decimal_precision(value: i128, precision: u8) -> Result<(), Arro
 /// interpreted as a Decimal number with precision `precision`
 #[inline]
 pub fn is_validate_decimal_precision(value: i128, precision: u8) -> bool {
-    let idx = usize::from(precision) - 1;
     precision <= DECIMAL128_MAX_PRECISION
-        && value >= MIN_DECIMAL_FOR_EACH_PRECISION[idx]
-        && value <= MAX_DECIMAL_FOR_EACH_PRECISION[idx]
+        && value >= MIN_DECIMAL_FOR_EACH_PRECISION[precision as usize]
+        && value <= MAX_DECIMAL_FOR_EACH_PRECISION[precision as usize]
 }
 
 /// Validates that the specified `i256` of value can be properly
@@ -773,16 +783,15 @@ pub fn validate_decimal256_precision(value: i256, precision: u8) -> Result<(), A
             "Max precision of a Decimal256 is {DECIMAL256_MAX_PRECISION}, but got {precision}",
         )));
     }
-    let idx = usize::from(precision) - 1;
-    if value > MAX_DECIMAL_BYTES_FOR_LARGER_EACH_PRECISION[idx] {
+    if value > MAX_DECIMAL_BYTES_FOR_LARGER_EACH_PRECISION[precision as usize] {
         Err(ArrowError::InvalidArgumentError(format!(
             "{value:?} is too large to store in a Decimal256 of precision {precision}. Max is {:?}",
-            MAX_DECIMAL_BYTES_FOR_LARGER_EACH_PRECISION[idx]
+            MAX_DECIMAL_BYTES_FOR_LARGER_EACH_PRECISION[precision as usize]
         )))
-    } else if value < MIN_DECIMAL_BYTES_FOR_LARGER_EACH_PRECISION[idx] {
+    } else if value < MIN_DECIMAL_BYTES_FOR_LARGER_EACH_PRECISION[precision as usize] {
         Err(ArrowError::InvalidArgumentError(format!(
             "{value:?} is too small to store in a Decimal256 of precision {precision}. Min is {:?}",
-            MIN_DECIMAL_BYTES_FOR_LARGER_EACH_PRECISION[idx]
+            MIN_DECIMAL_BYTES_FOR_LARGER_EACH_PRECISION[precision as usize]
         )))
     } else {
         Ok(())
@@ -793,8 +802,7 @@ pub fn validate_decimal256_precision(value: i256, precision: u8) -> Result<(), A
 /// interpreted as a Decimal number with precision `precision`
 #[inline]
 pub fn is_validate_decimal256_precision(value: i256, precision: u8) -> bool {
-    let idx = usize::from(precision) - 1;
     precision <= DECIMAL256_MAX_PRECISION
-        && value >= MIN_DECIMAL_BYTES_FOR_LARGER_EACH_PRECISION[idx]
-        && value <= MAX_DECIMAL_BYTES_FOR_LARGER_EACH_PRECISION[idx]
+        && value >= MIN_DECIMAL_BYTES_FOR_LARGER_EACH_PRECISION[precision as usize]
+        && value <= MAX_DECIMAL_BYTES_FOR_LARGER_EACH_PRECISION[precision as usize]
 }