@andygrove 2024-09-19
> thanks @andygrove probably a nit: I can see `usize::from(precision) - 1` a lot which require new allocation of usize. My feeling the precision is static? may we can precompute idx and pass it by reference?

What if we just insert an extra element in the start of each of the lookup arrays and then just use `precision as usize` directly as the index?