@andygrove 2024-09-19
Updated benchmark results after removing the `idx` allocation:

```
null_if_overflow_precision_128
                        time:   [12.555 µs 12.591 µs 12.628 µs]
                        change: [-85.358% -85.314% -85.275%] (p = 0.00 < 0.05)
                        Performance has improved.

null_if_overflow_precision_256
                        time:   [25.614 µs 25.755 µs 25.903 µs]
                        change: [-87.983% -87.827% -87.620%] (p = 0.00 < 0.05)
                        Performance has improved.

validate_decimal_precision_128
                        time:   [74.290 ns 74.703 ns 75.109 ns]
                        change: [-7.7409% -7.2664% -6.8248%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  5 (5.00%) low mild

validate_decimal_precision_256
                        time:   [215.27 ns 216.01 ns 216.75 ns]
                        change: [-4.4231% -3.7018% -2.9993%] (p = 0.00 < 0.05)
                        Performance has improved.
```