@andygrove 2024-09-20
> I'm okay with the PR, although the array that changed moved to crate visibility and doesn't seem to affect others, however I'm not sure if anyone else other than arrow-rs actually used it

I reverted the change to the original array and added a copy with the extra element so as not to break the current API. The new array is `pub(crate)`